use middleware admin and user

+++ routes/web.php
@@ -2,6 +2,7 @@
 
 use App\Http\Controllers\BlogPost;
 use App\Http\Controllers\BlogPostController;
+use App\Http\Controllers\HomeController;
 use App\Http\Controllers\ProfileController;
 use Illuminate\Support\Facades\Route;
 
@@ -10,21 +11,24 @@ Route::get('/', function () {
 });
 
 Route::middleware(['auth'])->group(function () {
-    Route::get('/dashboard/createBlogPost', [BlogPostController::class, 'createPost'])->name('createBlogPost');
-    Route::post('/dashboard/posts', [BlogPostController::class, 'store'])->name('posts.store');
-
     Route::get('/dashboard', [BlogPostController::class, 'showLatestBlog'])->name('dashboard');
-
-    Route::get('/dashboardPage', [BlogPostController::class, 'dashboardPage'])->name('profile.dashboardPage');
-    Route::get('/dashboardPage/edit/{id}', [BlogPost::class, 'edit'])->name('posts.edit');
-    Route::put('/dashboardPage/update/{id}', [BlogPost::class, 'update'])->name('posts.update');
-    Route::delete('/dashboardPage/{id}', [BlogPost::class, 'destroy'])->name('posts.destroy');
-
-
-
+    Route::get('/blogDetails/{id}', [BlogPost::class, 'blogDetails'])->name('blogDetails');
+    
     Route::get('/profile', [ProfileController::class, 'edit'])->name('profile.edit');
     Route::patch('/profile', [ProfileController::class, 'update'])->name('profile.update');
     Route::delete('/profile', [ProfileController::class, 'destroy'])->name('profile.destroy');
 });
 
 require __DIR__ . '/auth.php';
+
+
+Route::middleware(['auth', 'admin'])->group(function () {
+    Route::get('/admin/dashboard', [HomeController::class, 'showLatestBlog'])->name('admin.dashboard');
+    Route::get('/admin/dashboard/createBlogPost', [HomeController::class, 'createPost'])->name('createBlogPost');
+    Route::post('/admin/dashboard/posts', [HomeController::class, 'store'])->name('posts.store');
+    Route::get('/admin/dashboardPage', [HomeController::class, 'dashboardPage'])->name('profile.dashboardPage');
+    Route::get('/admin/dashboardPage/edit/{id}', [BlogPost::class, 'edit'])->name('posts.edit');
+    Route::put('/admin/dashboardPage/update/{id}', [BlogPost::class, 'update'])->name('posts.update');
+    Route::delete('/admin/dashboardPage/{id}', [BlogPost::class, 'destroy'])->name('posts.destroy');
+    // Route::get('admin/dashboard', [HomeController::class, 'showLatestBlog'])->name('dashboard');
+});
